backend_api/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.1.0
packaging==25.0
pluggy==1.6.0
psycopg[binary]==3.2.12
Pygments==2.19.2
pytest==9.0.1
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
PyJWT==2.8.0
flask-cors==5.0.0
